Quick and dirty conversion to individual ptrs to remember data store
This is a very quick and dirty conversion to lay down (bring back?) the groundwork to have individual pointers to each array of the tuple_vector data stored by a list of parents to the TupleVecImpl. The big part of figuring out the ptrs and storing them individually is not in yet, and so this only works with a tuple_vector<int> and there's some other things regarding allocation and moves that have to be fixed up but this is the innards of the core of what that'll be.