Check for empty range in vector::erase(first,last)
When erase() is called on a vector with an empty range (first == last),
it should be a no-op. If erase() does not check for this case, it can
end up calling the move assignment operator for an obect with itself as
the source and dest. This is unnecessary work at best, and at worst
could cause problems if the object's implementation of the move
assignment does not handle this case well, leaving the object in an
undefined state.

This fix simply adds a check for first==last and avoids doing any work
in that case.