Allowed kSize[] to be updated without requiring manual updates to kSizeCount or kArraySizeMax. (#170)
Various sort benchmarks now operate on a vector of sort functions instead of explicitly skipping sort functions in multiple locations. This change also eliminates the need to handle sort functions in switch statements in order to skip them.

The primary reason for these changes is so that it is easier to benchmark larger array sizes and/or to modify the sort functions that will be benchmarked (since it quickly becomes impractical to benchmark O(n^2) sort functions for larger array sizes).