Reworking of TupleVecIter to improve performance
Instead of storing a reference to the TupleVecImpl and querying that for the pointers to the data necessary when, e.g., resolving iterator contents, the iterator now stores a list of all of the TupleVecImpl's data pointers. It does increase the size of the Iterator, so copying it around is a bit more expensive, but resolving the iterator is much faster. Previously, the codegen would implement the two data hops to the TupleVecImpl then to the member for the data, as opposed to just the one data hop to the member for the data. Simple iterations are not affected, but transforms that involve a lot of dereferencing of iterators such as sorting or remove_if are much faster (e.g. the swap in eastl::remove_copy_if takes ~0.66x the time it did before)