Optimized radix_sort to reduce memory traffic and potentially skip some scatter operations depending on the input data being sorted.
This change also removes two specializations of radix_sort that can be handled by the generic code.
The change also introduces a template parameter for 'DigitBits' which defaults to 8.  In some cases users may be able to acheive even better performance by tuning this value based on the data being sorted.