Test for Dinkumware with _CPPLIB_VER instead of _YVALS.
In VS2017 "15.8" changes have been made to the standard library to allow
including some headers, like <type_traits>, in completely freestanding (not
linking with msvcp140.dll) mode, so that these may be consumed by the kernel
or similar constrained environments.

Unfortunately this causes problems because the link dependency on the DLL is
injected in yvals.h, so including <cstddef> no longer causes _YVALS to be
defined, since that is the header guard for the yvals.h header.

This file is already using _CPPLIB_VER for detection of specific Dinkumware
version, so it's already assuming _CPPLIB_VER indicates Dinkumware, and that
macro is at least pseudo-documented.