Adding tests and fixes for, e.g., comparison of tuples w/ non-const refs (#185)
"const" qualifier on const tuple_element_t<I, TupleImpl<Indices, Ts...>>& gets dropped when Ts is an lvalue ref. In this case, the only source of 'const-ness' on tuple_element_t comes from being baked into the source type. This means for non-const lvalue refs, the return type just bcomes "Ts&", but getInternal will attempt to return a const-ref values, and fail to compile because we cannot remove the const qualifier. Changing out the typename used by adding const specifically to the type of the ref, instead of trying to add it to the ref, solves the problem and we guarantee a const-ref return type.