Removing template customization from ctor overload
Need a better case to include it, really. It already has to be tuple_vector compatible, and the iterator can be implicitly cast if the tuple-types are compatible, tool. At this time, I'm not sure what case this would cover that isn't taken care of already (and can be done so automatically - e.g., reverse_iterator support needs to have their order swapped manually and can't be generically used)