Improved the sorting tests
The first improvement is to add a sanity check to make sure the sum of the array elements is the same after sorting as before sorting.  Otherwise a sort function that simply zeroed every element in the container could pass the 'is_sorted' test.

The second improvement was to add a test that verifies if the stable sort algorithms are actually stable.