Fix UB causing crashes with some compilers.
It's not allowed (Undefined Behavior) to call memmove with null pointers even if the size is 0.
In vector::DoInsertValueEnd memmove is called via eastl::uninitialized_move_ptr_if_noexcept unconditionally even for null pointers (when vector is empty).
Therefore, further checks for p in vector::DoFree and allocator::deallocate get optimized away which leads to crashes.