Add char8_t support for string_view's ""sv operators
Enables the ""sv and ""_sv operators if the compiler implements
support for the distinct char8_t type in present in C++20.

Tests are included which *should* trigger under the MSVC part of
the CI setup; I've verified locally that it all works with Clang 9
and GCC 9.2.0 with a local hack to compile only the string_view
tests in C++20 mode.

Includes a specialization of hash<u8string_view> to enable the
tests to actually compile successfully.