Improve char8_t support for strings
When char8_t support is enabled within a compiler, include a
literal ""s operator for converting a u8"" string literal to
an instance of u8string.

Enabled tests for u8string if char8_t is a distinct type, which
as with the string_view support required another hash specialization
to be defined, as well as a few extra overloads of DecodePart to
prevent compilation errors in tests that append strings of mixed
encodings. Each added overload simply casts the char8_t* data to
char* and defers to another overload to handle the operation.