Fix overriding of -std=c++17 flags (#342)
* Fix overriding of -std=c++17 flags

From local testing with other changes I'd noticed that C++17 features
like deduction guides weren't being properly enabled in tests. The
logic of the CommonCppFlags script causes the flags set for C++17
support to be overridden by C++14 flags instead due to the if()
check immediately after in each branch being a GREATER_OR_EQUALS
check.

* Remove disabled random_shuffle benchmarks

random_shuffle was deprecated and later removed as of C++17. The
benchmark was commented out, but the implementation remained and
triggers compilation errors under Clang 9.

* Fix inclusion of <variant> on OSX Clang

The CI setup seems to build with Xcode 9.4 which has a C++17
compatible Clang compiler but a libc++ that lacks <variant>.