Add EASTL_BUILD_CHAR8T_TESTS option to test build
This enables building a subset of the tests (currently string
and string_view) with char8_t support if the compiler being
used supports either the `-fchar8_t` or `/Zc:char8_t` flag.

If this option is enabled a diagostic will be printed if the
flag is successfully found or not.