Fail test compilation if char8_t support is missing
If the tests were compiled with char8_t support requested, and the
appropriate compiler flag was found, we set a define which can be
tested for to trigger a compile error if EA_CHAR8_UNIQUE still
evaluates to false.

This should work to prove that the tests are correctly compiling
with char8_t support enabled under the CI setup when the CMake
configure process says it's doing so.