Build benchmarks without char8_t support
Under recent versions of MSVC, the use of the `/permissive-` flag
in CommonCppFlags enables char8_t support which caused compilation
failures with the tests and benchmarks.

As this isn't supported yet, if the compiler supports a flag that
explicitly disables char8_t support (`-fno-char8_t` or `/Zc:char8_t-`)
then we compile both the tests and benchmarks with that flag by
default, overriding only for specific files that do support it.

Use of add_compile_options instead of `set(CMAKE_CXX_FLAGS ...)` is
to ensure the flag propagates down to the submodules, since for some
reason this wasn't happening reliably.