Limit inclusion of functional.h by utility.h
A circular dependency was forming when function_detail.h was
included ahead of function.h, which could occur if a user included
fixed_function.h prior to functional.h.

As utility.h only requires the type traits for
eastl::reference_wrapper, this limits the inclusion to only
attempting to pull them in instead of the entirety of functional.h.

This may break users who depend upon utility.h pulling in functional.h.