Fix tuple vector assign (#374)
* Improve tuple_vector assign test

There is a bug when using the assign overload with two pointers to
tuples. When the target tuple_vector has enough capacity but a lower
size than the input range, the elements at the begining of the input
range are repeated in the target tuple_vector.

The existing test did not cover this case as all the elements in the
input range were identical. This change now uses distinct elements for
all the input std::initializer_list s provided to the assign function,
highlighting the bug.

* Fix tuple_vector assign with a range of values

The unitialized part of the target tuple_vector is now initialized with
the corresponding range in the input range, instead of always using the
start of the range.

Co-authored-by: Thierry Laviron <thierry.laviron@grenoble-inp.org>