Remove usage of c++14 extension
The `_v` shorthand is only available since c++14. This was probably an
oversight since the rest of the file uses the more compatible
`::value` syntax.

Another related problem was that the wrong guard was used to detect
the availability of the extension. There seem to be some internal
confusion about what a variable template actually is.