Don't use atomic_thread_fence under TSAN
TSAN is one of the best tools for checking correctness of
concurrent programs out there, but sadly, it does not support
atomic_thread_fence and unlikely ever will.

But for good codegen on ARM we NEED atomic_thread_fence in
shared_ptr. Hence, we need to selectively use it only when
TSAN is disabled and fallback to an unconditional acq_rel
operation where the ARM acquire fence will be performed even when
we are not releaseing the last reference. This is pretty much
equivalent semantically, but avoids false-positive TSAN reports.